Added a lot of Client -> Server packets

+++ Lib K Relay/Networking/Packets/Client/ActivePetUpdateRequestPacket.cs
@@ -2,25 +2,25 @@
 {
     public class ActivePetUpdateRequestPacket : Packet
     {
-        public const int FollowPet = 1;
-        public const int UnfollowPet = 2;
-        public const int ReleasePet = 3;
+        public const byte FollowPet = 1;
+        public const byte UnfollowPet = 2;
+        public const byte ReleasePet = 3;
 
-        public int CommandId;
-        public uint PetId;
+        public byte CommandId;
+        public int PetId;
 
         public override PacketType Type => PacketType.ACTIVEPETUPDATEREQUEST;
 
         public override void Read(PacketReader r)
         {
             CommandId = r.ReadByte();
-            PetId = (uint)r.ReadInt32();
+            PetId = r.ReadInt32();
         }
 
         public override void Write(PacketWriter w)
         {
-            w.Write((byte)CommandId);
-            w.Write((int)PetId);
+            w.Write(CommandId);
+            w.Write(PetId);
         }
     }
 }
